+++ doc/libopenmpt/changelog.md
@@ -67,6 +67,8 @@ is just a high-level summary.
  *  [**New**] Support for GCC 7 has been restored.
  *  [**New**] Support for Clang 6 has been restored.
  *  [**New**] Support for Android NDK 18 has been restored.
+ *  [**New**] openmpt123: `--banner [0|1|2]` allows changing the openmpt123
+    banner style to hidden, shown, and verbose, respectively.
 
  *  [**Change**] xmp-openmpt: The Amiga resampler emulation is now enabled by
     default.